Ejemplos/Pruebas Luis/pygame_book_examples/código capítulo 9/requirements.txt
pygame==2.1.2
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.5
